Move newKernel out to start the epic creation.
