+++ src/notebook/actions.js
@@ -1,10 +1,12 @@
 import Immutable from 'immutable';
-
 import Rx from 'rxjs/Rx';
 
 import * as agendas from './agendas';
-import { launchKernel } from './api/kernel';
 import * as constants from './constants';
+
+import { newKernel as newKernelImp } from './not-yet-epics/kernelLaunch';
+
+export const newKernel = newKernelImp;
 
 const path = require('path');
 
@@ -21,37 +23,6 @@ export function setLanguageInfo(langInfo) {
     langInfo,
   };
 }
-
-export function newKernel(kernelSpecName, cwd) {
-  return () => Rx.Observable.create((subscriber) => {
-    launchKernel(kernelSpecName, { cwd })
-      .then(kc => {
-        const { channels, connectionFile, spawn } = kc;
-
-        // Listen to the execution status of the kernel
-        channels.iopub
-          .filter(msg => msg.header.msg_type === 'status')
-          .map(msg => msg.content.execution_state)
-          .subscribe(() => subscriber.next(setExecutionState('idle')));
-
-        agendas.acquireKernelInfo(channels)
-          .subscribe(action => {
-            subscriber.next(action);
-            subscriber.next(setExecutionState('idle'));
-          });
-
-        subscriber.next({
-          type: constants.NEW_KERNEL,
-          channels,
-          connectionFile,
-          spawn,
-          kernelSpecName,
-        });
-      })
-      .catch((err) => console.error(err));
-  });
-}
-
 
 export function setNotebook(nbData, filename) {
   const cwd = (filename && path.dirname(path.resolve(filename))) || process.cwd();
